+++ src/mainwindow.hpp
@@ -46,7 +46,6 @@ class FakeVimCommand;
 namespace Extensions
 {
 class CFTool;
-class ClangFormatter;
 struct CompanionData;
 } // namespace Extensions
 
@@ -115,7 +114,7 @@ class MainWindow : public QMainWindow
     void compileOnly();
     void runOnly();
     void compileAndRun();
-    void formatSource();
+    void formatSource(bool selectionOnly, bool logOnNoChange);
 
     void applyCompanion(const Extensions::CompanionData &data);
 
@@ -209,7 +208,6 @@ class MainWindow : public QMainWindow
     QString language;
     bool isLanguageSet = false;
 
-    Extensions::ClangFormatter *formatter = nullptr;
     Core::Compiler *compiler = nullptr;
     QVector<Core::Runner *> runner;
     Core::Checker *checker = nullptr;